refactor: service 모듈로 서비스 구성요소 이동
- Service, Dto, ServiceTest
