+++ service/build.gradle
@@ -3,6 +3,7 @@ jar { enabled = true }
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.flywaydb:flyway-core'
     implementation 'com.h2database:h2'
 }
