+++ service/src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -1,6 +1,5 @@
 package kitchenpos.menu.dto;
 
-import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.product.domain.Product;
 import org.springframework.util.CollectionUtils;
